Move s3 and s3_base fixtures to dae_conftests.py
The s3 fixture is used both by the vcf loader tests and the
GenomicResourceURLRepo tests so it needs to be in a location that is
shared between the two.
